pyproject.toml
[tool.poetry]
name = "discordpy-bot-template-with-commands-extension"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
replit = "^3.2.4"
auth = "^0.5.3"
"discord.py" = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
